requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
python-jose[cryptography]
pydantic
passlib
pydantic[email]